requirements.txt
streamlit==1.51.0
langchain==1.2.0
langchain-core==1.2.5
langchain-openai==1.1.0
langchain-huggingface==1.1.0
python-dotenv==1.1.1
requests==2.32.5